added downloading backup in Excel format
