+++ orchestrator_service/src/main/java/com/override/orchestrator_service/mapper/TransactionMapper.java
@@ -4,7 +4,6 @@ import com.override.dto.TransactionDTO;
 import com.override.dto.TransactionResponseDTO;
 import com.override.dto.constants.Type;
 import com.override.orchestrator_service.model.Transaction;
-import com.override.orchestrator_service.model.User;
 import com.override.orchestrator_service.util.NumericalUtils;
 import org.springframework.stereotype.Component;
 
@@ -40,14 +39,6 @@ public class TransactionMapper {
             builder.categoryName(transaction.getCategory().getName());
             builder.type(transaction.getCategory().getType());
         }
-        StringBuilder usersNames = new StringBuilder();
-        for (User user : transaction.getAccount().getUsers()) {
-            if (usersNames.length() > 0) {
-                usersNames.append(", ");
-            }
-            usersNames.append(user.getUsername());
-        }
-        builder.telegramUserName(usersNames.toString());
 
         return builder.build();
     }
