+++ orchestrator_service/src/main/resources/static/scripts/setting_scripts.js
@@ -69,7 +69,7 @@ function saveFile() {
 }
 function downloadExcel() {
     $.ajax({
-        url: './settings/backup/excel',
+        url: './settings/export/excel',
         method: 'GET',
         xhrFields: {
             responseType: 'blob'
