+++ orchestrator_service/src/main/resources/templates/settings.html
@@ -34,7 +34,7 @@
         <button id="downloadButton" class="get-backup-button">Скачать бекап данных json</button>
     </div>
     <div class="info">
-        <button id="downloadExelButton" class="get-backup-button">Скачать бекап данных Excel</button>
+        <button id="downloadExelButton" class="get-backup-button">Экспорт истории транзакций в Excel</button>
     </div>
 </div>
 <div class="info">
